requirements.txt
requests>=2.31.0
PyYAML>=6.0
Jinja2>=3.1.0
python-crontab>=2.7.0